requirements_backend.txt
google-generativeai
python-dotenv
pymongo
langchain
faiss-cpu
sentence-transformers
rank-bm25
langchain-ollama
langchain-community
langchain-huggingface
fastapi
uvicorn
ollama
python-multipart
selenium
webdriver-manager
ultralytics
langchain_google_genai
wikipedia
google
googlesearch-python
webserver